Organizer Conversions
Added refresh functions to main and tabs.
A 60 second timer now loops in main. When it finishes, it calls refresh on each tab and each tab then calls refresh on each module.
This refreshing uses the Data Organizer structure.
